✨ feat: useGetQuizDetailQuery를 통해서 데이터 페칭 구현
