+++ src/app/quiz/[quizId]/@recommendation/page.tsx
@@ -1,5 +1,5 @@
 import { QuizCarousel } from '@/app/quiz/components';
-import { getRecommendationByQuizId } from '@/app/quiz/remotes/quiz';
+import { getRecommendationByQuizId } from '@/app/quiz/remotes/recommendation';
 import { Text } from '@/common';
 
 type Props = {
