+++ src/app/quiz/models/quiz.ts
@@ -2,7 +2,6 @@ export type QuizType = 'A_B_IMAGE' | 'A_B_SIMPLE' | 'O_X_IMAGE' | 'O_X_SIMPLE';
 export type AB = 'A' | 'B';
 export type OX = 'O' | 'X';
 export type QuizButtonType = AB | OX;
-export type ButtonNumber = '1' | '2';
 
 export interface QuizDetailResponse {
   data: QuizResponse;
@@ -41,7 +40,7 @@ export interface Question {
   question: string;
   imageUrl?: string;
   buttons: {
-    [key in ButtonNumber]: QuizButton;
+    [key in QuizButtonType]: QuizButton;
   };
 }
